add separate travis job for tests that require multiple (non-nodejs) runtimes
